Rename ThemeContext to ThemeProvider and place it in Providers
