+++ src/providers/Providers.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 
 import { TooltipProvider } from '../common/components/ui/tooltip';
 
-import { ThemeProvider } from '@/contexts/ThemeContext';
 import { CollectionsProvider } from '@/providers/CollectionsProvider';
 import { InvocationProvider } from '@/providers/InvocationProvider';
+import { ThemeProvider } from '@/providers/ThemeProvider';
 
 const Providers = ({ children }: { children: React.ReactNode }) => {
 	return (
